Collection sharing

+++ routes/collections.js
@@ -5,60 +5,102 @@ const router = express.Router();
 const auth = require("../middleware/auth");
 const validateObjID = require("../middleware/validateObjID");
 const { User } = require("../models/user");
-const { Collection, validateCollection } = require("../models/collection");
+const { Collection } = require("../models/collection");
+const { CollectionData } = require("../models/collectionData");
+const { ActionToken } = require("../models/actionToken");
 const {
   CollectionModel,
   validateCollectionModel,
 } = require("../models/collectionModel");
-const { CollectionData } = require("../models/collectionData");
 const {
   CollectionSettings,
   validateCollectionSettings,
 } = require("../models/collectionSettings");
-const { ActionToken } = require("../models/actionToken");
 
-// ===Collection===
-router.get("/:id", [validateObjID, auth], async (req, res) => {
-  const userId = await getUserIdByEmail(req.user.email);
+async function getCollection(collectionId, user, populate = false) {
   let collection;
-  if (req.query.noPopulate) {
+  if (populate)
     collection = await Collection.findOne({
-      _id: req.params.id,
-      owner: userId,
-    });
-  } else {
-    collection = await Collection.findOne({
-      _id: req.params.id,
-      owner: userId,
+      _id: collectionId,
+      $or: [{ owner: user._id }, { "sharedTo.userEmail": user.email }],
     })
       .populate("model")
       .populate("data")
       .populate("settings");
+  else
+    collection = await Collection.findOne({
+      _id: collectionId,
+      $or: [{ owner: user._id }, { "sharedTo.userEmail": user.email }],
+    });
+  return collection;
+}
+
+function checkEditPermissions(collection, user) {
+  // User is the collection's owner
+  if (collection.owner.toString() === user._id.toString()) return true;
+
+  // User has the collection shared to him
+  const share = collection.sharedTo.find((x) => x.userEmail === user.email);
+  return share && share.role === "edit";
+}
+
+async function getUserByEmail(email, idOnly = true) {
+  let user = await User.findOne({
+    email: email,
+  });
+
+  if (idOnly) return user ? user._id : null;
+  else return user ? user : null;
+}
+
+// ===Collection===
+router.get("/:id", [validateObjID, auth], async (req, res) => {
+  let collection;
+  if (req.query.noPopulate) {
+    collection = (await getCollection(req.params.id, req.user)).toObject();
+  } else {
+    collection = (await getCollection(req.params.id, req.user, true)).toObject();
   }
+
   if (!collection) return res.status(404).send("Collection not found.");
-  else return res.status(200).send(collection);
+
+  // Change the "owner" value
+  let user = await User.findById(collection.owner);
+  collection.ownerString = `${user.name} (${user.email === req.user.email ? "you" : user.email})`;
+  return res.status(200).send(collection);
 });
 
 router.get("/", [auth], async (req, res) => {
-  const userId = await getUserIdByEmail(req.user.email);
-  const collections = await Collection.find({ owner: userId });
-  res.status(200).send(collections);
+  const user = await User.findById(req.user._id);
+  let collections = await Collection.find({
+    $and: [
+      { _id: { $in: user.collections } },
+      {
+        $or: [
+          { owner: user._id },
+          { sharedTo: { $elemMatch: { $eq: user._id } } },
+        ],
+      },
+    ],
+  });
+  return res.status(200).send(collections);
 });
 
 router.post("/", [auth], async (req, res) => {
   // Validate collection title
-  if (!(req.body.title && req.body.title.length <= 20))
-    return res
-      .status(400)
-      .send("Collection title can be min 1 and max 30 characters long.");
+  let { error } = Joi.object({
+    title: Joi.string().max(30).required(),
+  }).validate(req.body);
+  if (error) return res.status(400).send(error.details[0].message);
 
   // Prepare Collection data structure
   let collection = new Collection({ title: req.body.title });
 
   // Check and set ownership
-  let user_id = await getUserIdByEmail(req.user.email);
-  if (!user_id) return res.status(400).send("Creator does not exist.");
-  collection.owner = user_id;
+  let user = await getUserByEmail(req.user.email, false);
+  if (!user) return res.status(400).send("Creator does not exist.");
+  collection.owner = user._id; //
+  user.collections.push(collection._id);
 
   // Create and set an empty data record
   let collectionData = new CollectionData({
@@ -77,7 +119,7 @@ router.post("/", [auth], async (req, res) => {
   collection.settings[0] = collectionSettings._id;
 
   // Create and set collection model if it is valid
-  const { error } = validateCollectionModel(req.body.model, true);
+  error = validateCollectionModel(req.body.model, true).error;
   if (error) return res.status(400).send(error.details[0].message);
   let collectionModel = new CollectionModel({
     parent: collection._id,
@@ -86,8 +128,9 @@ router.post("/", [auth], async (req, res) => {
   collection.model = collectionModel._id;
   await collectionModel.save();
 
-  // Save the newly created collection
+  // Save the newly created collection and altered user
   await collection.save();
+  await User.findByIdAndUpdate(user._id, _.omit(user, ["__v", "_id"]));
 
   return res.status(201).send(collection);
 });
@@ -100,11 +143,7 @@ router.put("/:id", [validateObjID, auth], async (req, res) => {
   if (error) return res.status(400).send(error.details[0].message);
 
   // Find collection and check ownership
-  const userId = await getUserIdByEmail(req.user.email);
-  let collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  let collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
 
   collection.title = req.body.title;
@@ -119,17 +158,17 @@ router.put("/:id", [validateObjID, auth], async (req, res) => {
 router.delete("/:id", [validateObjID, auth], async (req, res) => {
   // First time user sends a collection DELETE request, he gets a confirmation token.
   // In order to delete the desired collection, he has to repeat the API call, together with the delete action token
-  let userId = await getUserIdByEmail(req.user.email);
+  let user = await getUserByEmail(req.user.email, false);
   let collection = await Collection.findOne({
     _id: req.params.id,
-    owner: userId,
+    owner: user._id,
   });
   if (!collection) return res.status(404).send("Collection not found.");
 
   if (!req.query.actionToken) {
     let deleteToken = new ActionToken({
       purpose: `Delete collection "${collection.title}"`,
-      userId: userId,
+      userId: user._id,
       targetId: collection._id,
     });
     await deleteToken.save();
@@ -137,13 +176,16 @@ router.delete("/:id", [validateObjID, auth], async (req, res) => {
   } else {
     let deleteToken = await ActionToken.findOne({
       _id: req.query.actionToken,
-      userId: userId,
+      userId: user._id,
       targetId: collection._id,
     });
     if (!deleteToken)
       return res
         .status(404)
         .send("No valid token found to approve this action.");
+    const index = user.collections.indexOf(collection._id);
+    user.collections.splice(index, 1);
+    await User.findByIdAndUpdate(user._id, user);
     await Collection.findByIdAndDelete(deleteToken.targetId);
     await CollectionModel.findOneAndDelete({ parent: deleteToken.targetId });
     await CollectionData.findOneAndDelete({ parent: deleteToken.targetId });
@@ -156,11 +198,7 @@ router.delete("/:id", [validateObjID, auth], async (req, res) => {
 // ===CollectionModel===
 router.get("/:id/model", [validateObjID, auth], async (req, res) => {
   // Get the parent (collection)
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
 
   // Get the collection model
@@ -175,11 +213,7 @@ router.get("/:id/model", [validateObjID, auth], async (req, res) => {
 // ===CollectionData===
 router.get("/:id/data", [validateObjID, auth], async (req, res) => {
   // Get the parent (collection)
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
 
   // Get the collection data
@@ -193,12 +227,14 @@ router.get("/:id/data", [validateObjID, auth], async (req, res) => {
 
 router.post("/:id/data", [validateObjID, auth], async (req, res) => {
   // Get the parent collection
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
+
+  // Check for edit permissions
+  if (!checkEditPermissions(collection, req.user))
+    return res
+      .status(403)
+      .send("You are not permitted to edit the collection.");
 
   // Get the collection data
   let collectionData = await CollectionData.findOne({
@@ -216,17 +252,19 @@ router.post("/:id/data", [validateObjID, auth], async (req, res) => {
     _.omit(collectionData, ["_id", "__v"]),
     { new: true }
   );
-  res.status(200).send("Added new row.");
+  return res.status(200).send("Added new row.");
 });
 
 router.put("/:id/data/:index", [validateObjID, auth], async (req, res) => {
   // Get the parent collection
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
+
+  // Check for edit permissions
+  if (!checkEditPermissions(collection, req.user))
+    return res
+      .status(403)
+      .send("You are not permitted to edit the collection.");
 
   // Get the collection data
   let collectionData = await CollectionData.findOne({
@@ -243,17 +281,19 @@ router.put("/:id/data/:index", [validateObjID, auth], async (req, res) => {
     _.omit(collectionData, ["_id", "__v"]),
     { new: true }
   );
-  res.status(200).send(`Edited row at index ${req.params.index}`);
+  return res.status(200).send(`Edited row at index ${req.params.index}`);
 });
 
 router.delete("/:id/data/:index", [validateObjID, auth], async (req, res) => {
   // Get the parent collection
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
+
+  // Check for edit permissions
+  if (!checkEditPermissions(collection, req.user))
+    return res
+      .status(403)
+      .send("You are not permitted to edit the collection.");
 
   // Get the collection data
   let collectionData = await CollectionData.findOne({
@@ -271,17 +311,13 @@ router.delete("/:id/data/:index", [validateObjID, auth], async (req, res) => {
     _.omit(collectionData, ["_id", "__v"]),
     { new: true }
   );
-  res.status(200).send(`Removed row at index ${req.params.index}`);
+  return res.status(200).send(`Removed row at index ${req.params.index}`);
 });
 
 // ===CollectionSettings===
 router.get("/:id/settings", [validateObjID, auth], async (req, res) => {
   // Get the parent (collection)
-  const userId = await getUserIdByEmail(req.user.email);
-  const collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  const collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
 
   // Get the collection settings
@@ -295,12 +331,14 @@ router.get("/:id/settings", [validateObjID, auth], async (req, res) => {
 
 router.post("/:id/settings/new", [validateObjID, auth], async (req, res) => {
   // Get the parent (collection)
-  const userId = await getUserIdByEmail(req.user.email);
-  let collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  let collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
+
+  // Check for edit permissions
+  if (!checkEditPermissions(collection, req.user))
+    return res
+      .status(403)
+      .send("You are not permitted to edit the collection.");
 
   // Create new, save it and edit the parent collection object
   let newSettings = new CollectionSettings({
@@ -317,12 +355,14 @@ router.post("/:id/settings/new", [validateObjID, auth], async (req, res) => {
 
 router.put("/:id/settings/reset", [validateObjID, auth], async (req, res) => {
   // Get the parent (collection)
-  const userId = await getUserIdByEmail(req.user.email);
-  let collection = await Collection.findOne({
-    _id: req.params.id,
-    owner: userId,
-  });
+  let collection = await getCollection(req.params.id, req.user);
   if (!collection) return res.status(404).send("Collection not found.");
+
+  // Check for edit permissions
+  if (!checkEditPermissions(collection, req.user))
+    return res
+      .status(403)
+      .send("You are not permitted to edit the collection.");
 
   // Delete the old settings object
   await CollectionSettings.deleteMany({ parent: collection._id });
@@ -349,12 +389,14 @@ router.put(
     const { error } = validateCollectionSettings(req.body);
     if (error) return res.status(400).send(error.details[0].message);
 
+    // Check for edit permissions
+    if (!checkEditPermissions(collection, req.user))
+      return res
+        .status(403)
+        .send("You are not permitted to edit the collection.");
+
     // Get the parent (collection)
-    const userId = await getUserIdByEmail(req.user.email);
-    const collection = await Collection.findOne({
-      _id: req.params.id,
-      owner: userId,
-    });
+    const collection = await getCollection(req.params.id, req.user);
     if (!collection) return res.status(404).send("Collection not found.");
 
     // Save the changes
@@ -374,12 +416,14 @@ router.patch(
   [validateObjID, auth],
   async (req, res) => {
     // Get the parent (collection)
-    const userId = await getUserIdByEmail(req.user.email);
-    let collection = await Collection.findOne({
-      _id: req.params.id,
-      owner: userId,
-    });
+    let collection = await getCollection(req.params.id, req.user);
     if (!collection) return res.status(404).send("Collection not found.");
+
+    // Check for edit permissions
+    if (!checkEditPermissions(collection, req.user))
+      return res
+        .status(403)
+        .send("You are not permitted to edit the collection.");
 
     // Validate the new name
     const payloadSchema = Joi.object({
@@ -399,18 +443,26 @@ router.patch(
 );
 
 router.delete(
-  "/:id/settings/:settings_id", [validateObjID, auth], async (req, res) => {
+  "/:id/settings/:settings_id",
+  [validateObjID, auth],
+  async (req, res) => {
     // Get the parent (collection)
-    const userId = await getUserIdByEmail(req.user.email);
-    let collection = await Collection.findOne({
-      _id: req.params.id,
-      owner: userId,
-    });
+    let collection = await getCollection(req.params.id, req.user);
     if (!collection) return res.status(404).send("Collection not found.");
 
+    // Check for edit permissions
+    if (!checkEditPermissions(collection, req.user))
+      return res
+        .status(403)
+        .send("You are not permitted to edit the collection.");
+
     // // Delete the collection settings
-    let collectionSettings = await CollectionSettings.findOneAndDelete({_id: req.params.settings_id, parent: req.params.id});
-    if (!collectionSettings) return res.status(404).send("Collection settings preset not found.");
+    let collectionSettings = await CollectionSettings.findOneAndDelete({
+      _id: req.params.settings_id,
+      parent: req.params.id,
+    });
+    if (!collectionSettings)
+      return res.status(404).send("Collection settings preset not found.");
 
     // Remove the collection settings from collection object
     for (let s in collection.settings)
@@ -418,17 +470,12 @@ router.delete(
         collection.settings.splice(s, 1);
         break;
       }
-    await Collection.findByIdAndUpdate(collection._id, _.omit(collection, ["_id, __v"]));
+    await Collection.findByIdAndUpdate(
+      collection._id,
+      _.omit(collection, ["_id, __v"])
+    );
     return res.status(200).send("Settings preset deleted.");
   }
-)
-
-async function getUserIdByEmail(email) {
-  let user = await User.findOne({
-    email: email,
-  });
-
-  return user ? user._id : null;
-}
+);
 
 module.exports = router;
